Remove review step and finalize immediately
