+++ docs/PlannedDesign.md
@@ -6,7 +6,7 @@
 
 ## 0. 目的 / スコープ
 - 目的：患者が最小限の入力で問診を完了し、LLMの追加質問で不足情報を補完できるUIを提供する。
-- スコープ：患者向け画面（基本情報＋受診種別→ベース問診→LLM追加質問→最終確認→完了）、および管理画面（問診テンプレート設定、LLM接続設定、簡易プレビュー）。
+ - スコープ：患者向け画面（基本情報＋受診種別→ベース問診→LLM追加質問→完了）、および管理画面（問診テンプレート設定、LLM接続設定、簡易プレビュー）。
 - 非スコープ：禁忌/赤旗、辞書正規化、同意画面、重複照合/ID発行、ルールベース追質問、緊急分岐。
 
 ---
@@ -21,13 +21,11 @@
 3. **/questionnaire**
    - ベース問診フォーム（テンプレートに従う）
    - 送信後、LLMの追加質問フェーズへ遷移
-4. **/questions**
-   - 追加質問（LLM生成）を 1問ずつ or 複数まとめて表示・回答
-   - 追加質問がなくなった/上限到達/手動終了で次へ
-5. **/review**
-   - 全回答の一覧表示・その場修正・確定
-6. **/done**
-   - 完了表示（印刷/共有は任意）
+  4. **/questions**
+     - 追加質問（LLM生成）を 1問ずつ or 複数まとめて表示・回答
+     - 追加質問がなくなった/上限到達/手動終了で完了へ
+  5. **/done**
+     - 完了表示（印刷/共有は任意）
 
 ### 管理者向け
 1. **/admin/login**
@@ -49,7 +47,7 @@
 - **ヘッダー**：左にロゴ（MonshinMate）、右に「管理画面」ボタン（常時表示、/adminへ。/admin はテンプレート一覧へリダイレクト）。
 - **フッター**：`本システムはローカルLLMを使用しており、外部へ情報が送信されることはありません。` を小さめに常時表示。
 - **ブレッドクラム**：患者向けは非表示（ステップナビで代替）。管理画面のみ簡易パンくず。
-- **ステップナビ（患者向け）**：Entry → 問診 → 追加質問 → 確認 → 完了（現在位置を強調）。
+ - **ステップナビ（患者向け）**：Entry → 問診 → 追加質問 → 完了（現在位置を強調）。
 
 ---
 
@@ -88,12 +86,7 @@
 - **状態**：進行インジケータ（例：2/3 質問）
 - **終了条件**：LLM結果が0件／上限到達／ユーザー手動終了
 
-### 3.5 Review（/review）
-- **内容**：全回答（ベース問診＋追加質問）を一覧表示、各項目は**その場編集**可
-- **アクション**：確定送信（/done へ）／戻る（/questions）
-- **UI**：セクションごとに編集リンク、編集はインライン
-
-### 3.6 Done（/done）
+### 3.5 Done（/done）
 - **表示**：完了メッセージ、受付に提示するための要約（テキスト/Markdown）
 - **任意**：印刷、共有（クリップボードコピー）
 
@@ -157,10 +150,7 @@
   - 入力：`questions: LlmQuestion[]`
   - 送信時に `{ id, answer }[]` を返す
 
-### 5.4 レビュー/要約
-- `ReviewTable`
-  - 入力：すべての回答
-  - インライン編集対応
+### 5.4 完了/要約
 - `SummaryBlock`
   - 入力：回答
   - 出力：印刷用テキスト/Markdown（フロント側で整形）
@@ -176,7 +166,6 @@
 - **ガード**：
   - `/questionnaire` は `session_id` が未作成なら `/` へ（Entry にて作成）
   - `/questions` は問診送信済が前提
-  - `/review` は LLM 追加質問フェーズ完了が前提（またはスキップ）
   - 管理配下は認証必須（未認証は `/admin/login` へ）
 
 ---
@@ -267,7 +256,7 @@
 1) ルーティング骨格 & 共通レイアウト
 2) Entry / VisitType / Questionnaire
 3) LLM 追加質問画面（モーダル or リスト）
-4) Review / Done
+4) Done
 5) 管理：Login / Templates / Sessions / LLM
 6) ドラフト保存・再試行キューなどの改善
