+++ docs/implementation.md
@@ -8,7 +8,7 @@
 
 ## 0. 目的 / スコープ
 - 目的：
-  - 患者名・生年月日の入力 → 初診/再診分岐 → テンプレートに基づくベース問診 → LLMによる最小追加質問 → 最終確認 → 完了、までを **細い導線** で実装する。
+  - 患者名・生年月日の入力 → 初診/再診分岐 → テンプレートに基づくベース問診 → LLMによる最小追加質問 → 完了、までを **細い導線** で実装する。
 - スコープ：
   - 患者向けフローと管理画面（テンプレート管理、LLM接続設定）
   - バックエンドAPI（セッション、テンプレートCRUD、LLM質問生成、最終化）
@@ -47,7 +47,7 @@
 ### MS4：要約と保存
 - 目的：全回答の一覧確認→確定→保存/要約出力
 - **完了条件**
-  - [x] `/review` で全回答を一望しインライン編集可能
+  - [x] 追加質問終了後、自動的にセッションを確定して要約を生成
   - [x] `POST /sessions/{id}/finalize` が `summaryText` と `allAnswers` を返す
   - [x] 完了ステータスとタイムスタンプ保存
   - [x] LLM 設定が有効かつ疎通OKの場合、バックグラウンドでカスタムプロンプトを用いたサマリー生成を実行（UIは非ブロッキング）
@@ -56,12 +56,12 @@
 ### MS5：フロント実装（患者/管理）
 - 目的：UI導線と状態管理の実装
 - **完了条件**
-  - [x] **患者**：Entry（氏名/生年月日＋受診種別）→ Questionnaire → Questions → Review → Done
+  - [x] **患者**：Entry（氏名/生年月日＋受診種別）→ Questionnaire → Questions → Done
     - [x] Entry で氏名・生年月日を入力
     - [x] Entry で「当院の受診は初めてですか？」を選択（「初めて」= initial / 「受診したことがある」= followup）しセッション作成
     - [x] Questionnaire で text/multi/yesno/date に応じた入力フォームを表示
     - [x] Questions で追加質問を順次表示
-    - [x] Review で回答一覧を表示しインライン編集後確定へ進む
+    - [x] 追加質問終了後は自動的に完了画面へ遷移
   - [x] Done で要約を表示
   - [x] **共通**：ヘッダー右上「管理画面」ボタン、フッター注意文の常時表示
 - [x] **管理**：ログイン→テンプレート→問診結果→LLM設定
@@ -122,7 +122,6 @@
    - [x] `/`：氏名・生年月日＋受診種別の入力（選択後にセッション作成→`/questionnaire` へ）
    - [x] `/questionnaire`：テンプレに基づくフォーム（軽量条件表示、ドラフト保存）
    - [x] `/questions`：LLM追加質問（モーダル or カード列）と進行インジケータ
-   - [x] `/review`：全回答の一覧・インライン編集・確定
    - [x] `/done`：完了メッセージと要約（印刷/コピー任意）
 3) **管理向け**
    - [x] `/admin/login`：管理者ログイン
@@ -195,7 +194,7 @@
 ## 9. 受け入れ基準（抜粋）
 - ベース問診のみでもセッションを完了できる
 - 追加質問は上限N件を超えない / 項目ごとの再質問は最大3回
-- 最終確認画面で全入力を一望し、その場編集後に確定できる
+- 完了画面では回答一覧の編集は行わない
 - スタッフ向け要約が生成され、印刷/EMR貼付に足る簡潔さ
 - UI上に **「ローカルLLM利用・外部送信なし」** が恒常表示
 
@@ -281,8 +280,8 @@
 - [x] LLM 設定が無効または疎通テストに失敗している場合は、管理画面で追質問の設定をオンにできないようにした。
 
 ## 23. LLM フォローアップ判定待機画面の追加（2025-10-31）
-- [x] 固定フォーム回答後、追質問の有無を判定する間に待機画面を表示するようにした。
-- [x] 判定結果に応じて追加質問画面または確認画面へ自動遷移する。
+ - [x] 固定フォーム回答後、追質問の有無を判定する間に待機画面を表示するようにした。
+ - [x] 判定結果に応じて追加質問画面または完了画面へ自動遷移する。
 
 ## 24. 管理/チャット画面の左右余白を削減（2025-08-23）
 - [x] 管理画面のコンテナを全幅化し、左右パディングを縮小。
@@ -495,3 +494,7 @@
 - [x] パスワード送信後の二段階認証コード入力も同じモーダル内で行えるようにした。
 - [x] 変更: `frontend/src/pages/AdminLogin.tsx`。
 - [x] ドキュメント更新: `docs/admin_system_setup.md`, `frontend/public/docs/admin_system_setup.md`。
+
+## 61. 問診完了後の確認画面を廃止（2025-11-17）
+- [x] Review 画面を削除し、質問終了後は直接完了画面へ遷移するよう変更。
+- [x] フロントエンドおよびドキュメントを更新。
