+++ docs/plannedSystem.md
@@ -5,7 +5,7 @@
 ### 1. 方針・前提
 - 本実装は、本書の問診ロジック仕様を最小構成で実装する計画である。
 - **本システムはローカルLLMを使用しており、外部へ情報が送信されることはありません。**（UI上にも明記する）
-- スコープに含む：患者名・生年月日の取得、初診/再診の分岐、テンプレートに基づくベース問診、LLMによる不足項目の**必要最小限**の追加質問、最終確認・確定、保存、簡易要約。
+ - スコープに含む：患者名・生年月日の取得、初診/再診の分岐、テンプレートに基づくベース問診、LLMによる不足項目の**必要最小限**の追加質問、確定、保存、簡易要約。
 - スコープ外：禁忌・赤旗、辞書/正規化、緊急分岐、重複照合/新規ID発行、同意画面、ルールベース追質問（ハードガード）。
 
 #### 1.1 非機能要件 / 設計方針（最小）
@@ -54,7 +54,7 @@
 - **MS2 セッション基盤**：`SessionFSM` 初期化（`remaining_items`/`completion_status`/`attempt_counts`）、`StructuredContextManager` による永続更新。
 - **MS3 追加質問ループ**：`step("answer")` で回答受領→`LLMGateway.generate_followups` により回答全体を基に追加質問を生成→`Validator.is_complete` 判定→不足時に追質問（項目ごと最大3回・セッションあたりの上限N）。
 - **MS4 要約と保存**：必要情報が揃ったら `_finalize_item` を実行し、`collected_data` と `SessionResponse` に反映。全項目完了時にセッション終了。
-- **MS5 フロントエンド**：患者フロー（基本情報＋受診種別→ベース問診→一次送信→追加質問→最終確認→完了）、管理UI（テンプレ編集）。
+ - **MS5 フロントエンド**：患者フロー（基本情報＋受診種別→ベース問診→一次送信→追加質問→完了）、管理UI（テンプレ編集）。
 - **MS6 ログ/観測性（最小）**：セッション進行ログ、LLM I/O メタ（トークン/所要時間）保存。PII は最小化し、必要に応じマスキング。
 - **MS7 UAT/受け入れ**：想定シナリオでエンドツーエンド検証、受け入れ基準を満たせばリリース。
 
@@ -69,7 +69,7 @@
 3) **API（例）**：
    - `POST /sessions`：患者名・生年月日・受診種別・`questionnaire_id` を受け新規セッション生成。
    - `POST /sessions/{id}/answer`：`{item_id, answer}` を受け、次の質問（ベース or 追加）を返す。必要に応じて `questions[]` の配列で複数返却可。
-   - `POST /sessions/{id}/finalize`：最終確認後の確定。要約（テキスト/Markdown）と `collected_data` を返す。
+     - `POST /sessions/{id}/finalize`：回答送信後の確定。要約（テキスト/Markdown）と `collected_data` を返す。
    - `GET /questionnaires/{id}/template?visit_type=initial|followup`：対象テンプレートの取得。
    - `GET /llm/settings`：現在のLLM設定を取得。
    - `PUT /llm/settings`：LLM設定を更新。
@@ -84,8 +84,7 @@
    - 画面A：基本情報（患者名・生年月日＋受診種別）
    - 画面C：ベース問診フォーム（テンプレートに沿った入力）
    - 画面D：一次送信→LLM追加質問の提示と回答（1問ずつ or 複数）
-   - 画面E：最終確認（全回答の一覧/その場修正）→確定
-   - 画面F：完了（受付向け番号や案内は任意）
+     - 画面E：完了（受付向け番号や案内は任意）
 2) **管理フロー**：問診テンプレ CRUD、テンプレプレビュー（初診/再診）。
 3) **UI 表示文**：フッター等に「本システムはローカルLLMを使用しており、外部へ情報が送信されることはありません。」を常時表示。
 
@@ -196,6 +195,6 @@
 ### 9. 受け入れ基準（抜粋）
 - ベース問診のみでもセッションを完了できる。
 - 追加質問は上限 N 件を超えない。項目ごとの再質問は最大 3 回。
-- 最終確認画面で全入力が一望でき、その場で修正して確定できる。
+- 完了画面では回答一覧の編集は行わない。
 - スタッフ向け要約が生成され、印刷/EMR への貼付に足る簡潔さである。
 - UI 上に「ローカルLLM利用・外部送信なし」の明記が恒常表示されている。
