+++ docs/session_api.md
@@ -39,7 +39,7 @@
 - **備考**: 型や選択肢を検証し、不正な場合は 400 を返す。空欄で送信された回答は「該当なし」として保存される。
 
 ## POST /sessions/{session_id}/finalize
-- **概要**: セッションを確定し要約を生成する。
+- **概要**: セッションを確定し要約を生成する（回答送信後すぐに呼び出される）。
 - **レスポンス**:
   - `summary` (str): 生成された要約。サマリー作成モードが無効の場合は空文字
   - `answers` (object): 確定した回答
