+++ frontend/src/pages/LlmWait.tsx
@@ -1,11 +1,26 @@
 import { useEffect } from 'react';
 import { VStack, Spinner, Text } from '@chakra-ui/react';
 import { useNavigate } from 'react-router-dom';
+import { postWithRetry } from '../retryQueue';
 
 /** LLM 追質問の要否判定待機画面。 */
 export default function LlmWait() {
   const navigate = useNavigate();
   const sessionId = sessionStorage.getItem('session_id');
+
+  const finalize = async () => {
+    if (!sessionId) return;
+    try {
+      const res = await fetch(`/sessions/${sessionId}/finalize`, { method: 'POST' });
+      const data = await res.json();
+      sessionStorage.setItem('summary', data.summary);
+    } catch (e) {
+      console.error('finalize failed', e);
+      postWithRetry(`/sessions/${sessionId}/finalize`, {});
+      alert('ネットワークエラーが発生しました。接続後に再度お試しください。');
+    }
+    navigate('/done');
+  };
 
   useEffect(() => {
     if (!sessionId) {
@@ -20,12 +35,11 @@ export default function LlmWait() {
         if (data.questions && data.questions.length > 0) {
           navigate('/questions');
         } else {
-          navigate('/review');
+          await finalize();
         }
       } catch (e) {
         console.error('llm question check failed', e);
-        // 失敗時は追質問をスキップ
-        navigate('/review');
+        await finalize();
       }
     };
     check();
